Data type bug fixed in spike analysis view

+++ src/engine/RecordingManager.h
@@ -41,8 +41,8 @@ struct MetadataChunk;
 struct SpikeTrain {
 	SpikeTrain() : upperThresh(0), lowerThresh(0), channelIndex(0), color(0) {}
 	std::vector<int64_t> spikes;
-	int16_t upperThresh;
-	int16_t lowerThresh;
+	int upperThresh;
+	int lowerThresh;
     int16_t channelIndex;
 
 	int color;
